#31 Updating class/package names due to refactoring in Phileas.
